Split BlobService from FileService.
