+++ deepwell/src/services/blob/structs.rs
@@ -1,5 +1,5 @@
 /*
- * services/file/mod.rs
+ * services/blob/structs.rs
  *
  * DEEPWELL - Wikijump API provider and database manager
  * Copyright (C) 2019-2022 Wikijump Team
@@ -18,13 +18,4 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-mod prelude {
-    pub use super::super::prelude::*;
-    pub use super::structs::*;
-}
-
-mod service;
-mod structs;
-
-pub use self::service::FileService;
-pub use self::structs::*;
+// TODO
