+++ deepwell/src/services/file/service.rs
@@ -19,8 +19,7 @@
  */
 
 use super::prelude::*;
-use crate::hash::{hash_to_hex, sha512_hash, Hash};
-use std::str;
+use crate::models::file::{self, Entity as File, Model as FileModel};
 
 #[derive(Debug)]
 pub struct FileService;
@@ -30,7 +29,7 @@ impl FileService {
     ///
     /// In the background, this stores the blob via content addressing,
     /// meaning that duplicates are not uploaded twice.
-    pub async fn create(ctx: &ServiceContext<'_>) -> Result<()> {
+    pub async fn create(ctx: &ServiceContext<'_>, data: &[u8]) -> Result<()> {
         // TODO insert into file
 
         todo!()
@@ -54,39 +53,6 @@ impl FileService {
         todo!()
     }
 
-    /// Hard deletes this file and all duplicates.
-    ///
-    /// This is a very powerful method and needs to be used carefully.
-    /// It should only be accessible to platform staff.
-    ///
-    /// As opposed to normal soft deletions, this method will completely
-    /// remove a file from Wikijump. The file rows will be deleted themselves,
-    /// and will cascade to any places where file IDs are used.
-    ///
-    /// This method should only be used very rarely to clear content such
-    /// as severe copyright violations, abuse content, or comply with court orders.
-    pub async fn hard_delete_all(ctx: &ServiceContext<'_>, file_id: &str) -> Result<()> {
-        // TODO delete blob, all used files
-        // TODO add to audit log
-
-        // Delete blob from S3
-        //
-        // This needs to be last, since unlike the other steps,
-        // this cannot be rolled back by the database.
-        let bucket = ctx.s3_bucket();
-        let hash: Hash = todo!();
-        let hex_hash = hash_to_hex(&hash);
-
-        let (_, status) = bucket.delete_object(&hex_hash).await?;
-
-        match status {
-            204 => (),
-            _ => return s3_error(&[], status, "hard-deleting S3 blob"),
-        }
-
-        todo!()
-    }
-
     /// Gets an uploaded file that has been, including its contents if requested.
     pub async fn get_optional(
         ctx: &ServiceContext<'_>,
@@ -104,78 +70,22 @@ impl FileService {
         }
     }
 
-    /// Creates a blob with this data, if it does not already exist.
-    pub async fn create_blob(ctx: &ServiceContext<'_>, data: &[u8]) -> Result<Hash> {
-        let bucket = ctx.s3_bucket();
-        let hash = sha512_hash(data);
-        let hex_hash = hash_to_hex(&hash);
+    /// Hard deletes this file and all duplicates.
+    ///
+    /// This is a very powerful method and needs to be used carefully.
+    /// It should only be accessible to platform staff.
+    ///
+    /// As opposed to normal soft deletions, this method will completely
+    /// remove a file from Wikijump. The file rows will be deleted themselves,
+    /// and will cascade to any places where file IDs are used.
+    ///
+    /// This method should only be used very rarely to clear content such
+    /// as severe copyright violations, abuse content, or comply with court orders.
+    pub async fn hard_delete_all(ctx: &ServiceContext<'_>, file_id: &str) -> Result<()> {
+        // TODO find hash. update all files with the same hash
+        // TODO add to audit log
+        // TODO hard delete BlobService
 
-        // TODO insert into file_blob table
-
-        let (return_data, status) = bucket.put_object(&hex_hash, data).await?;
-
-        // We assume all unexpected statuses are errors, even if 1XX or 2XX
-        match status {
-            200 => Ok(hash),
-            _ => s3_error(&return_data, status, "creating S3 blob"),
-        }
+        todo!()
     }
-
-    pub async fn get_blob_optional(
-        ctx: &ServiceContext<'_>,
-        hash: &[u8],
-    ) -> Result<Option<Vec<u8>>> {
-        let bucket = ctx.s3_bucket();
-        let hex_hash = hash_to_hex(hash);
-        let (data, status) = bucket.get_object(&hex_hash).await?;
-
-        // TODO read from file_blob table
-        // TODO change return to FileBlob type
-
-        match status {
-            200 => Ok(Some(data)),
-            404 => Ok(None),
-            _ => s3_error(&data, status, "fetching S3 blob"),
-        }
-    }
-
-    pub async fn blob_exists(ctx: &ServiceContext<'_>, hash: &[u8]) -> Result<bool> {
-        let bucket = ctx.s3_bucket();
-        let hex_hash = hash_to_hex(hash);
-        let (data, status) = bucket.get_object(&hex_hash).await?;
-
-        // TODO check from file_blob table, don't call S3
-
-        match status {
-            200 | 204 => Ok(true),
-            404 => Ok(false),
-            _ => s3_error(&data, status, "fetching S3 blob"),
-        }
-    }
-
-    #[inline]
-    pub async fn get_blob(ctx: &ServiceContext<'_>, hash: &[u8]) -> Result<Vec<u8>> {
-        match Self::get_blob_optional(ctx, hash).await? {
-            Some(string) => Ok(string),
-            None => Err(Error::NotFound),
-        }
-    }
-}
-
-/// Helper method to parse out an S3 error response and print the message (if any)
-fn s3_error<T>(data: &[u8], status: u16, action: &str) -> Result<T> {
-    let error_message = match str::from_utf8(data) {
-        Ok(m) if m.is_empty() => "(no content)",
-        Ok(m) => m,
-        Err(_) => "(invalid UTF-8)",
-    };
-
-    tide::log::error!(
-        "Error while {} (HTTP {}): {}",
-        action,
-        status,
-        error_message,
-    );
-
-    Err(Error::RemoteOperationFailed)
 }
